added messageDAO and conversationDAO, change messageID to long

+++ src/main/java/edu/cmu/db/entities/Message.java
@@ -14,7 +14,7 @@ import java.time.Instant;
 public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int messageID;
+    private long messageID;
 
     @Column(name = "messageContent")
     private String messageContent;
@@ -38,11 +38,11 @@ public class Message {
         this.conversationID = conversationID;
     }
 
-    public int getMessageID() {
+    public long getMessageID() {
         return messageID;
     }
 
-    public void setMessageID(int messageID) {
+    public void setMessageID(long messageID) {
         this.messageID = messageID;
     }
