+++ src/main/resources/db/migration/V35__create_message_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE message (
-  messageID INT NOT NULL AUTO_INCREMENT,
+  messageID      LONG NOT NULL AUTO_INCREMENT,
   PRIMARY KEY (messageID),
   messageContent VARCHAR(256),
-  startingTime DATETIME,
-  messageSender VARCHAR(256),
-  conversationID       INT,
+  startingTime   DATETIME,
+  messageSender  VARCHAR(256),
+  conversationID INT,
   FOREIGN KEY (conversationID) REFERENCES conversation (conversationID)
 );
